+++ InputGlyphs/Assets/InputGlyphs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.2.0] - 2024-09-08
+### Added
+- Added the Index value for Single Layout.
+### Changed
+- Changed the Image to be transformed using LayoutElement.
+### Modified
+- Modified to allow multiple components to be edited.
+
 ## [1.1.2] - 2024-09-02
 ### Added
 - Added shift key to texture map.